코인프로젝트1/requirements.txt
# ===== Core Dependencies =====
openai==1.11.0
requests==2.31.0
pyupbit==0.2.33
python-dotenv==1.0.1
streamlit==1.31.0
schedule==1.2.0
python-telegram-bot==21.0.1
plotly==5.18.0
pandas==2.2.0
numpy==1.26.4
apscheduler==3.10.4
PyYAML==6.0.1
SQLAlchemy==2.0.25
pytest==8.0.0
beautifulsoup4==4.12.3
lxml==5.1.0
yfinance==0.2.36
ta==0.11.0
scikit-learn==1.4.0

# ===== Additional for Production =====
tenacity==8.2.3       # 재시도 로직
python-dateutil==2.8.2
pytz==2024.1
tqdm==4.66.1         # 진행률 표시
loguru==0.7.2        # 향상된 로깅

# ===== Optional (uncomment if needed) =====
# ib_insync==0.9.86  # Interactive Brokers
# ccxt==4.2.34       # 다중 거래소 지원